renamed DepositActions to Interact
