--- src/renderer/components/depositActions/index.ts
@@ -1,1 +1,0 @@
-export * from './DepositActions'
