+++ src/renderer/components/interact/index.ts
@@ -1,0 +1,1 @@
+export * from './Interact'
