--- src/renderer/views/wallet/DepositActions/index.ts
@@ -1,1 +1,0 @@
-export * from './DepositActionsView'
