+++ src/renderer/views/wallet/Interact/InteractView.tsx
@@ -3,15 +3,15 @@ import React from 'react'
 import { Col, Row } from 'antd'
 import { useParams } from 'react-router'
 
-import { DepositActions } from '../../../components/depositActions'
+import { Interact } from '../../../components/interact'
 import { BackLink } from '../../../components/uielements/backLink'
 import * as walletRoutes from '../../../routes/wallet'
 import { BondView } from './BondView'
-import * as Styled from './DepositActionsView.styles'
+import * as Styled from './InteractView.styles'
 import { LeaveView } from './LeaveView'
 import { UnbondView } from './UnbondView'
 
-export const DepositActionView: React.FC = () => {
+export const InteractView: React.FC = () => {
   const { walletAddress } = useParams<walletRoutes.DepositParams>()
 
   return (
@@ -22,7 +22,7 @@ export const DepositActionView: React.FC = () => {
         </Col>
       </Row>
       <Styled.ContentContainer>
-        <DepositActions
+        <Interact
           bondContent={<BondView walletAddress={walletAddress} />}
           leaveContent={<LeaveView />}
           unbondContent={<UnbondView />}
