+++ src/renderer/views/wallet/Interact/index.ts
@@ -1,0 +1,1 @@
+export * from './InteractView'
